🧬 Migrate the base objects or variables to module *api_config._base*.
