+++ pymock_api/model/api_config/_base.py
@@ -1,0 +1,86 @@
+from abc import ABC, ABCMeta, abstractmethod
+from dataclasses import dataclass, field
+from typing import Any, Callable, Dict, Optional
+
+# The truly semantically is more near like following:
+#
+# ConfigType = TypeVar("ConfigType" bound="_Config")
+# def need_implement_func(param: ConfigType):
+#     ... (implementation)
+#
+# However, it would have mypy issue:
+# error: Argument 1 of "{method}" is incompatible with supertype "_Config"; supertype defines the argument type as
+# "ConfigType"  [override]
+# note: This violates the Liskov substitution principle
+# note: See https://mypy.readthedocs.io/en/stable/common_issues.html#incompatible-overrides
+SelfType = Any
+
+
+class _Config(metaclass=ABCMeta):
+    def __eq__(self, other: SelfType) -> bool:
+        if other is None:
+            return False
+        if not isinstance(other, self.__class__):
+            raise TypeError(
+                f"Cannot run equal operation between these 2 objects because of their data types is different. Be "
+                f"operated object: {type(self)}, another object: {type(other)}."
+            )
+        return self._compare(other)
+
+    @abstractmethod
+    def _compare(self, other: SelfType) -> bool:
+        pass
+
+    @abstractmethod
+    def serialize(self, data: Optional[SelfType] = None) -> Optional[Dict[str, Any]]:
+        pass
+
+    @staticmethod
+    def _ensure_process_with_not_empty_value(function: Callable) -> Callable:
+        def _(self, data: Dict[str, Any]) -> Optional[SelfType]:
+            if not data:
+                return data
+            return function(self, data)
+
+        return _
+
+    @abstractmethod
+    def deserialize(self, data: Dict[str, Any]) -> Optional[SelfType]:
+        pass
+
+    def _get_prop(self, data: Optional[object], prop: str) -> Any:
+        if not hasattr(data, prop) and not hasattr(self, prop):
+            raise AttributeError(f"Cannot find attribute {prop} in objects {data} or {self}.")
+        return (getattr(data, prop) if data else None) or getattr(self, prop)
+
+
+@dataclass(eq=False)
+class _TemplatableConfig(_Config, ABC):
+    apply_template_props: bool = True
+
+    # The settings which could be set by section *template* or override the values
+    base_file_path: str = "./"
+    config_path: str = field(default_factory=str)
+    config_path_format: str = field(default_factory=str)
+
+    _has_apply_template_props_in_config: bool = False
+
+    def _compare(self, other: SelfType) -> bool:
+        return self.apply_template_props == other.apply_template_props
+
+    def serialize(self, data: Optional[SelfType] = None) -> Optional[Dict[str, Any]]:
+        apply_template_props: bool = self._get_prop(data, prop="apply_template_props")
+        if self._has_apply_template_props_in_config:
+            return {
+                "apply_template_props": apply_template_props,
+            }
+        else:
+            return {}
+
+    @_Config._ensure_process_with_not_empty_value
+    def deserialize(self, data: Dict[str, Any]) -> Optional[SelfType]:
+        apply_template_props = data.get("apply_template_props", None)
+        if apply_template_props is not None:
+            self._has_apply_template_props_in_config = True
+            self.apply_template_props = apply_template_props
+        return self
